Use WCL API for spell info
The Blizzard API is missing a lot of spells. This is a bug reported on their forums months ago. Blizzard does not seem invested in their API as it continues to be an unreliable source. WCL's API is one of the few remaining options.

Spells appear to have globally unique ids, so we can just use one API for Shadowlands and TBC.